+++ hemera_udf/uniswap_v4/abi/uniswapv4_abi.py
@@ -306,4 +306,4 @@ HOOK_CALLED_EVENT = Event(
         "name": "HookCalled",
         "type": "event",
     }
-) 
+)