+++ hemera_udf/uniswap_v4/models/feature_uniswap_v4_pool_current_prices.py
@@ -34,4 +34,4 @@ class UniswapV4PoolCurrentPrices(HemeraModel):
                 "update_strategy": None,
                 "converter": general_converter,
             },
-        ] 
+        ]