+++ hemera_udf/uniswap_v4/models/feature_uniswap_v4_pool_prices.py
@@ -34,4 +34,4 @@ class UniswapV4PoolPrices(HemeraModel):
                 "update_strategy": None,
                 "converter": general_converter,
             },
-        ] 
+        ]